+++ CHANGELOG.md
@@ -7,6 +7,7 @@ These are the changes made to the Ky3M project from the beginning:
 ### Features
 
 - Recast unknown versions processing.
+- Add RELEASE and PUNISH bad saved ID handling.
 
 ## [Release 0.2.1](https://github.com/Procybit/Ky3m/releases/tag/v0.2.1) (2022-08-17) - v0.2.0 Important patches